Change the interface
