--- src/Interfaces/fonts.ts
@@ -1,7 +1,0 @@
-export enum FontSizeName {
-  XS = 'XS',
-  S = 'S',
-  M = 'M',
-  L = 'L',
-  XL = 'XL',
-}
