+++ src/styles.ts
@@ -2,7 +2,7 @@ export const Colors = {
   brand: '#9f250a',
   accent: '#53d585',
   grey200: '#c8c8c8',
-  grey: '#8e8e8e',
+  grey500: '#8e8e8e',
   grey800: '#44484B',
   grey900: '#2f2f2f',
 }
