+++ src/utils.tsx
@@ -1,6 +1,3 @@
-// Uteles
-import state from '@state'
-
 export function uuid() {
   // return ('' + 1e7 + -1e3 + -4e3 + -8e3 + -1e11).replace(/[10]/g, function() {
   return ('' + 1e7).replace(/[10]/g, function() {
